rolled useEffect into custom hook
